+++ app/src/js/reducers/workflows.js
@@ -3,7 +3,9 @@ import { set } from 'object-path';
 import {
   WORKFLOWS,
   WORKFLOWS_INFLIGHT,
-  WORKFLOWS_ERROR
+  WORKFLOWS_ERROR,
+  SEARCH_WORKFLOWS,
+  CLEAR_WORKFLOWS_SEARCH
 } from '../actions/types';
 import { createReducer } from '@reduxjs/toolkit';
 
@@ -11,11 +13,11 @@ export const initialState = {
   list: {
     data: [],
     meta: {},
-    params: {},
     inflight: false,
-    error: false
+    error: false,
   },
-  map: {}
+  map: {},
+  searchString: null,
 };
 
 function createMap (data) {
@@ -26,20 +28,41 @@ function createMap (data) {
   return map;
 }
 
+/**
+ * Filters intput data by filterstring on the data'a object's name.
+*
+ * @param {Array} rawData - An array of objects with a name parameter.
+ * @param {string} filterString - a string to check if name includes.
+ * @returns {Array} Filtered Array of rawData's objects who's names include filterString.
+ */
+export const filterData = (rawData, filterString) => {
+  if (filterString !== null) {
+    return rawData.filter(d => d.name.toLowerCase().includes(filterString.toLowerCase()));
+  }
+  return rawData;
+};
+
 export default createReducer(initialState, {
   [WORKFLOWS]: (state, action) => {
-    const { data } = action;
+    const { data: rawData } = action;
+    const data = filterData(rawData, state.searchString);
     set(state, 'map', createMap(data));
     set(state, ['list', 'data'], data);
     set(state, ['list', 'meta'], { queriedAt: Date.now() });
     set(state, ['list', 'inflight'], false);
     set(state, ['list', 'error'], false);
   },
-  [WORKFLOWS_INFLIGHT]: (state, action) => {
+  [WORKFLOWS_INFLIGHT]: (state) => {
     set(state, ['list', 'inflight'], true);
   },
   [WORKFLOWS_ERROR]: (state, action) => {
     set(state, ['list', 'inflight'], false);
     set(state, ['list', 'error'], action.error);
-  }
+  },
+  [SEARCH_WORKFLOWS]: (state, action) => {
+    set(state, ['searchString'], action.searchString);
+  },
+  [CLEAR_WORKFLOWS_SEARCH]: (state) => {
+    set(state, ['searchString'], null);
+  },
 });